CutsceneFix: change ScreenPercentage calculation to work with seperate CVar instead.
Game seems to be scaling it by ScreenPercentage anyway, so my code that also scales it by ScreenPercentage is redundant (and probably makes it get upscaled twice...)
Luckily the code can be salvaged and used for scaling seperately of ScreenPercentage, eg. if you just want cutscenes to have increased res. without affecting rest of the game.